practica-play-time/requirements.txt
absl-py==2.1.0
click==8.1.8
dlib==19.24.6
face-recognition==1.3.0
face_recognition_models==0.3.0
grpcio==1.69.0
Markdown==3.7
MarkupSafe==3.0.2
numpy==2.2.1
packaging==24.2
pillow==11.1.0
protobuf==5.29.3
python-dotenv==1.0.1
setuptools==75.8.0
six==1.17.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
watchdog==6.0.0
Werkzeug==3.1.3
